feat: Add docs addon on storybook
