--- packages/margaret/.storybook/addons.js
@@ -1,5 +1,0 @@
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-options/register';
