--- packages/margaret/.storybook/config.js
@@ -1,9 +1,0 @@
-import { configure } from '@storybook/react';
-
-const req = require.context('../stories', true, /\.stories\.js$/);
-
-function loadStories() {
-  req.keys().forEach(filename => req(filename));
-}
-
-configure(loadStories, module);
