+++ packages/margaret/.storybook/main.js
@@ -1,0 +1,9 @@
+module.exports = {
+  stories: ['../src/components/*.stories.mdx', '../src/**/*.stories.js'],
+  addons: [
+    '@storybook/addon-docs',
+    '@storybook/addon-a11y',
+    '@storybook/addon-actions',
+    '@storybook/addon-knobs',
+  ],
+};
