+++ packages/margaret/package.json
@@ -2,20 +2,27 @@
   "name": "@tymate/margaret",
   "version": "0.1.1",
   "private": false,
+  "main": "./dist/index.js",
   "repository": "git@github.com:tymate/frontend-bricks.git",
   "license": "Apache-2.0",
   "dependencies": {
+    "@storybook/addon-docs": "^5.3.9",
     "@storybook/addon-info": "^5.1.8",
     "loaders.css": "^0.1.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "styled-components": "^4.3.1"
+    "prop-types": "^15.7.2",
+    "react-modal": "^3.11.1"
+  },
+  "peerDependencies": {
+    "react": ">= 16.12.0",
+    "react-dom": ">= 16.12.0",
+    "react-router-dom": ">= 5.X",
+    "styled-components": ">= 4.4"
   },
   "scripts": {
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "build": "babel src --out-dir dist"
+    "build": "babel src --out-dir dist",
+    "dev": "nodemon --watch src --exec 'yarn build'"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -26,13 +33,18 @@
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-stage-2": "^7.0.0",
-    "@storybook/addon-a11y": "^5.1.8",
-    "@storybook/addon-actions": "^5.1.8",
-    "@storybook/addon-knobs": "^5.1.8",
-    "@storybook/addon-links": "^5.1.8",
-    "@storybook/addon-options": "^5.1.8",
-    "@storybook/addons": "^5.1.8",
-    "@storybook/react": "^5.1.8",
-    "babel-loader": "^8.0.6"
+    "@storybook/addon-a11y": "^5.3.9",
+    "@storybook/addon-actions": "^5.3.9",
+    "@storybook/addon-knobs": "^5.3.9",
+    "@storybook/addon-links": "^5.3.9",
+    "@storybook/addon-options": "^5.3.9",
+    "@storybook/addons": "^5.3.9",
+    "@storybook/react": "^5.3.9",
+    "babel-loader": "^8.0.6",
+    "babel-watch": "^7.0.0",
+    "nodemon": "^2.0.2",
+    "react-is": "^16.12.0",
+    "react-router-dom": "^5.1.2",
+    "styled-components": "^5.0.0"
   }
 }
