+++ packages/margaret/src/components/basics.stories.mdx
@@ -1,0 +1,80 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { boolean } from '@storybook/addon-knobs';
+import { action } from '@storybook/addon-actions';
+import Button from './Button';
+import { Buttons } from '../ui'
+
+<Meta title="Button" component={Button} />
+
+<Props />
+
+# Button
+
+With `MDX` we can define a story for `Checkbox` right in the middle of our
+markdown documentation.
+
+## Default
+<Preview>
+  <Story name="DefaultButton">
+    <Buttons alignment="left">
+      <Button 
+        onClick={action('clicked')}
+        disabled={boolean('Disabled', false)}
+        isLoading={boolean('Loading', false)}
+      >
+        Hello! Button
+      </Button>
+      <Button disabled onClick={action('clicked')} >
+        Hello Button
+      </Button>
+    </Buttons>
+  </Story>
+</Preview>
+
+## Primary
+<Preview>
+  <Story name="PrimaryButton">
+    <Buttons alignment="left">
+      <Button
+        variant="primary"
+        disabled={boolean('Disabled', false)}
+        isLoading={boolean('Loading', false)}
+        onClick={action('clicked')}
+      >
+        Hello Button
+      </Button>
+    </Buttons>
+  </Story>
+</Preview>
+
+## Secondary
+<Preview>
+  <Story name="SecondaryButton">
+    <Buttons alignment="left">
+      <Button
+        variant="secondary"
+        disabled={boolean('Disabled', false)}
+        isLoading={boolean('Loading', false)}
+        onClick={action('clicked')}
+      >
+        Hello Button
+      </Button>
+    </Buttons>
+  </Story>
+</Preview>
+
+## Outline
+<Preview>
+  <Story name="OutlineButton">
+    <Buttons alignment="left">
+      <Button
+        variant="outline"
+        disabled={boolean('Disabled', false)}
+        isLoading={boolean('Loading', false)}
+        onClick={action('clicked')}
+      >
+        Hello Button
+      </Button>
+    </Buttons>
+  </Story>
+</Preview>
