+++ packages/margaret/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as Button } from './Button';
+export { default as MargaretProvider } from './MargaretProvider';
+export { default as Spinner } from './Spinner';
