+++ packages/margaret/src/index.js
@@ -1,5 +1,1 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { Button } from './ui';
-
-render(<Button>Test</Button>, document.querySelector('#root'));
+export * from './components';
