+++ packages/margaret/src/ui/theme.js
@@ -1,0 +1,25 @@
+import { spacing } from './utils';
+
+export const theme = {
+  primary: 'hsl(137, 54%, 61%)',
+  primaryLight: 'hsl(137, 54%, 70%)',
+  secondary: 'hsl(196, 53%, 41%)',
+  secondaryLight: 'hsl(196, 53%, 50%)',
+
+  disabled: 'rgba(4%, 6%, 13%, .2)',
+  separator: 'rgba(4%, 6%, 13%, .2)',
+
+  text: '#0B0F21',
+  textLight: 'rgba(4%, 6%, 13%, .5)',
+
+  spacing,
+
+  borderRadius: '6px',
+  buttonBorderRadius: '6px',
+
+  fontStack: {
+    title: '"Archivo Narrow", sans-serif',
+    body: '"Chivo", sans-serif',
+    ui: 'sans-serif',
+  },
+};
