+++ packages/margaret/stories/index.stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { ThemeProvider, createGlobalStyle } from 'styled-components';
+import { theme } from '../src/ui';
 import { addDecorator } from '@storybook/react';
 import { withA11y } from '@storybook/addon-a11y';
 import { withKnobs } from '@storybook/addon-knobs';
@@ -24,31 +25,7 @@ const OurStyles = createGlobalStyle`
 `;
 
 const ThemeDecorator = storyFn => (
-  <ThemeProvider
-    theme={{
-      primary: 'hsl(137, 54%, 61%)',
-      primaryLight: 'hsl(137, 54%, 70%)',
-      secondary: 'hsl(196, 53%, 41%)',
-      secondaryLight: 'hsl(196, 53%, 50%)',
-
-      disabled: 'rgba(4%, 6%, 13%, .2)',
-      separator: 'rgba(4%, 6%, 13%, .2)',
-
-      text: '#0B0F21',
-      textLight: 'rgba(4%, 6%, 13%, .5)',
-
-      spacing: (input = 1) => `${input * 16}px`,
-
-      borderRadius: '6px',
-      buttonBorderRadius: '6px',
-
-      fontStack: {
-        title: '"Archivo Narrow", sans-serif',
-        body: '"Chivo", sans-serif',
-        ui: 'sans-serif',
-      },
-    }}
-  >
+  <ThemeProvider theme={theme}>
     <React.Fragment>
       <OurStyles />
